requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Timezone support
pytz>=2023.3

# OpenAI
openai>=1.0.0

# Telegram
python-telegram-bot>=20.0

# Google Calendar
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Testing
pytest>=7.4.0

